updated information about tests and the organization of tests as well as updating the dependency for browser info
