+++ src/test/file-name-heading.test.ts
@@ -1,0 +1,23 @@
+import dedent from 'ts-dedent';
+import {rulesDict} from '../rules';
+
+describe('File Name Heading', () => {
+  it('Handles stray dashes', () => {
+    const before = dedent`
+      Text 1
+
+      ---
+      
+      Text 2
+        `;
+    const after = dedent`
+      # File Name
+      Text 1
+      
+      ---
+      
+      Text 2
+        `;
+    expect(rulesDict['file-name-heading'].apply(before, {'metadata: file name': 'File Name'})).toBe(after);
+  });
+});
