+++ src/test/missing-fields.test.ts
@@ -1,0 +1,9 @@
+import {rules, Rule} from '../rules';
+
+describe('Check missing fields', () => {
+  test.each(rules)('$name', (rule: Rule) => {
+    expect(rule.name).toBeTruthy();
+    expect(rule.description).toBeTruthy();
+    expect(rule.examples.length).toBeGreaterThan(0);
+  });
+});
