+++ src/test/remove-multiple-spaces.test.ts
@@ -1,0 +1,55 @@
+import dedent from 'ts-dedent';
+import {rulesDict} from '../rules';
+
+describe('Remove Multiple Spaces', () => {
+  it('Make sure spaces at the end of the line are ignored', () => {
+    const before = dedent`
+      # Hello world
+  
+      Paragraph contents are here  
+      Second paragraph contents here  
+      `;
+
+    const after = dedent`
+      # Hello world
+  
+      Paragraph contents are here  
+      Second paragraph contents here  
+      `;
+
+    expect(rulesDict['remove-multiple-spaces'].apply(before)).toBe(after);
+  });
+  it('Make sure spaces at the start of the line are ignored', () => {
+    const before = '  Paragraph contents are here\n  Second paragraph here...';
+
+    const after = '  Paragraph contents are here\n  Second paragraph here...';
+
+    expect(rulesDict['remove-multiple-spaces'].apply(before)).toBe(after);
+  });
+  it('Make sure non-letter followed or preceeded by 2 spaces has them cut down to 1 space', () => {
+    const before = dedent`
+    # Hello world
+
+    Paragraph contents  (something). Something else  .
+    `;
+
+    const after = dedent`
+    # Hello world
+
+    Paragraph contents (something). Something else .
+    `;
+
+    expect(rulesDict['remove-multiple-spaces'].apply(before)).toBe(after);
+  });
+  it('Links followed by parentheses do not prevent other removal of multiple spaces in a row', () => {
+    const before = dedent`
+      [Link text](path/fileName.md) (2 spaces in between  text)
+      `;
+
+    const after = dedent`
+      [Link text](path/fileName.md) (2 spaces in between text)
+      `;
+
+    expect(rulesDict['remove-multiple-spaces'].apply(before)).toBe(after);
+  });
+});
